pre-installation/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-security-checker"
version = "0.1.0"
description = "Pre-installation security checker for AI/ML models"
readme = "docs/RULESET.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "typer[all]>=0.9",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "jsonschema>=4.0",
]

[project.scripts]
checker = "checker.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
